Make the spotlights dynamic.
Only the most recent spotlight from each spotlight category will
appear on the home page in this order: People, Firms, Titles. The home
page will only display the title and excerpt. Users will need to
follow the link on the title to get to the full post.

fixes sfu-dhil/wphp#104