Person name search
The problem here is that the first name and last name fields are different. So when searching for a person by name we must search the first name and last name fields.

So I have changed the person name search so that including quotes around a name will change the nature of searching from a full text match to a first name last name comparison.

Examples:
 - `"charlotte smith"` will find the one Charlotte Smith in the database.
 - `charlotte smith` will find all the charlottes and all the smiths
 - `"rachel biggs"` will not find Rachel Charlotte Biggs because the search  term must include Charlotte.
 - `"rachel%biggs"` will find Rachel Charlotte Biggs. % is a wildcard  character for this one specific case.

fixes sfu-dhil/wphp#223