Rewrite ESTC imprint search.
Similar to how the person name search works, if you include quotes then
the quoted string must be present in the result.

`"a. james"` the search results will filter out anything not containing
that string, including punctuation. The filter is case insensitive,
so `"A. JAMES"` will return the same result.

`a. james` searches exactly the same as before.

fixes sfu-dhil/wphp#222