Don't renumber collection items when removing items.
Fixes sfu-dhil/wphp#284

I thought the problem was in Symfony, but no. It's in the javascript
that manages the collection UI. It was renumbering collection items when
removing them. Doctrine was interpreting the result incorrectly.

Given collection with items A[1] B[2] C[3] D[4], removing B would
result in A[1], C[2], D[3] and then doctrine would flush these changes
by updating 2 to C, resulting in A[1] C[2] C[2] D[3] which is nonsense
and prevented by unique keys. With this change, removing B results in
A[1] C[3] D[4] and doctrine does the right thing of only removing B.